third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/block-mixed-content-hides-warning-expected.txt

CONSOLE ERROR: Loading the script 'http://127.0.0.1:8080/security/contentSecurityPolicy/resources/alert-fail.js' violates the following Content Security Policy directive: "default-src 'self'". Note that 'script-src-elem' was not explicitly set, so 'default-src' is used as a fallback. The action has been blocked.
This page should neither alert "FAIL" nor generate any mixed content warnings in the console. 
